requirements.txt
streamlit
tensorflow
numpy
pandas
pillow
opencv-python-headless
reportlab